pyproject.toml
[project]
name = "law-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "openai>=2.8.1",
    "pdfplumber>=0.11.8",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
